+++ manager/server/internal/model/entity/quota_flow.go
@@ -8,7 +8,7 @@ package entity
 type QuotaFlow struct {
 	Id          int    `json:"id"          orm:"id"           ` //
 	Quota       int    `json:"quota"       orm:"quota"        ` //
-	Usage       int    `json:"usage"       orm:"usage"        ` //
+	Usage       int64  `json:"usage"       orm:"usage"        ` //
 	Target      string `json:"target"      orm:"target"       ` //
 	TargetType  string `json:"targetType"  orm:"target_type"  ` //
 	RecordStart int64  `json:"recordStart" orm:"record_start" ` //
